Make it runnable on windows diff --git a/grid-world/.gitignore b/grid-world/.gitignore new file mode 100644 index 0000000..d5e4f15 --- /dev/null +++ b/grid-world/.gitignore @@ -0,0 +1 @@ +**/__pycache__ \ No newline at end of file diff --git a/grid-world/graph/graph_utils.py b/grid-world/graph/graph_utils.py index 4a260b5..7f564fe 100644 --- a/grid-world/graph/graph_utils.py +++ b/grid-world/graph/graph_utils.py @@ -39,9 +39,9 @@ def batch_bin_encode(bin_tensor):      # bin_tensor: Nbatch x dim      if type(bin_tensor)==torch.Tensor:          if bin_tensor.dim()==2: -            return torch.mv(bin_tensor.type(torch.long), torch.from_numpy(1 << np.arange(bin_tensor.shape[-1])) ) +            return torch.mv(bin_tensor.type(torch.int32), torch.from_numpy(1 << np.arange(bin_tensor.shape[-1])) )          else: -            return torch.dot(bin_tensor.type(torch.long), torch.from_numpy(1 << np.arange(bin_tensor.shape[-1])) ).item() +            return torch.dot(bin_tensor.type(torch.int32), torch.from_numpy(1 << np.arange(bin_tensor.shape[-1])) ).item()      elif type(bin_tensor)==np.ndarray:          return bin_tensor.dot(1 << np.arange(bin_tensor.shape[-1]))      else:

+++ grid-world/graph/graph_utils.py
@@ -39,9 +39,9 @@ def batch_bin_encode(bin_tensor):
     # bin_tensor: Nbatch x dim
     if type(bin_tensor)==torch.Tensor:
         if bin_tensor.dim()==2:
-            return torch.mv(bin_tensor.type(torch.long), torch.from_numpy(1 << np.arange(bin_tensor.shape[-1])) )
+            return torch.mv(bin_tensor.type(torch.int32), torch.from_numpy(1 << np.arange(bin_tensor.shape[-1])) )
         else:
-            return torch.dot(bin_tensor.type(torch.long), torch.from_numpy(1 << np.arange(bin_tensor.shape[-1])) ).item()
+            return torch.dot(bin_tensor.type(torch.int32), torch.from_numpy(1 << np.arange(bin_tensor.shape[-1])) ).item()
     elif type(bin_tensor)==np.ndarray:
         return bin_tensor.dot(1 << np.arange(bin_tensor.shape[-1]))
     else:
